requirement.txt
fastapi==0.115.0
uvicorn[standard]==0.22.0
asyncpg==0.30.0
sqlalchemy[asyncio]>= 2.0.32
pydantic==2.9.2
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator==2.2.0
aioredis==2.0.1
